+++ mle_toolbox/launch/multi_config.py
@@ -1,4 +1,5 @@
 import logging
+from typing import Union
 from ..job import spawn_multiple_configs
 
 
@@ -7,6 +8,7 @@ def run_multiple_configs(
     meta_job_args: dict,
     single_job_args: dict,
     multi_config_args: dict,
+    message_id: Union[str, None] = None,
 ):
     """Run an experiment over different configurations (+random seeds)."""
     # 1. Create multiple experiment instances and submit the jobs
@@ -18,5 +20,6 @@ def run_multiple_configs(
         meta_job_args["experiment_dir"],
         num_seeds=multi_config_args["num_seeds"],
         random_seeds=multi_config_args["random_seeds"],
+        message_id=message_id,
         logger_level=logging.INFO,
     )