+++ mle_toolbox/utils/__init__.py
@@ -33,5 +33,5 @@ __all__ = [
     "load_hyper_log",
     "print_framed",
     "get_jax_os_ready",
-    "save_pkl_object"
+    "save_pkl_object",
 ]