+++ mle_toolbox/utils/core_files_load.py
@@ -28,6 +28,8 @@ def load_mle_toolbox_config(config_fname: str = "~/mle_config.toml") -> DotMap:
                     "use_venv_virtual_env": False,
                     "use_credential_encryption": False,
                     "random_seed": 42,
+                    "use_gcloud_protocol_sync": False,
+                    "local_protocol_fname": "~/local_mle_protocol.db",
                 }
             }
         )